requirements.txt
# pytorch: we suggest installing separately using the instructions at https://pytorch.org/
mlflow
pandas
numpy
pyro-ppl==1.6.0
